Add timeline to CollectiveOverview
