+++ components/dashboard/sections/Overview/IndividualOverview.tsx
@@ -1,79 +1,28 @@
 import React from 'react';
-import { useQuery } from '@apollo/client';
-import { flatten } from 'lodash';
 import { Megaphone, X } from 'lucide-react';
 import { useRouter } from 'next/router';
-import { FormattedMessage, useIntl } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import { HELP_MESSAGE } from '../../../../lib/constants/dismissable-help-message';
-import { API_V2_CONTEXT } from '../../../../lib/graphql/helpers';
-import type { WorkspaceHomeQuery } from '../../../../lib/graphql/types/v2/graphql';
-import { ActivityClassesI18N } from '../../../../lib/i18n/activities-classes';
 
 import DismissibleMessage from '../../../DismissibleMessage';
-import ExpenseDrawer from '../../../expenses/ExpenseDrawer';
 import { FEEDBACK_KEY, FeedbackModal } from '../../../FeedbackModal';
-import { Flex } from '../../../Grid';
 import Image from '../../../Image';
-import MessageBox from '../../../MessageBox';
-import MessageBoxGraphqlError from '../../../MessageBoxGraphqlError';
-import StyledButton from '../../../StyledButton';
-import { makeTruncatedValueAllSelectedLabelContainer, StyledSelectFilter } from '../../../StyledSelectFilter';
-import { H2 } from '../../../Text';
 import { Alert, AlertDescription, AlertTitle } from '../../../ui/Alert';
 import { Button } from '../../../ui/Button';
 import DashboardHeader from '../../DashboardHeader';
 import { DashboardSectionProps } from '../../types';
 
-import { workspaceHomeQuery } from './queries';
-import TimelineItem from './TimelineItem';
-const PAGE_SIZE = 20;
-
-const REACT_SELECT_COMPONENT_OVERRIDE = {
-  ValueContainer: makeTruncatedValueAllSelectedLabelContainer(
-    <FormattedMessage id="Dashboard.AllActivities" defaultMessage="All activities" />,
-  ),
-  MultiValue: () => null, // Items will be displayed as a truncated string in `TruncatedValueContainer `
-};
-
-const getFilterOptions = intl => [
-  { value: 'EXPENSES,VIRTUAL_CARDS', label: intl.formatMessage(ActivityClassesI18N['expenses.title']) },
-  { value: 'CONTRIBUTIONS', label: intl.formatMessage(ActivityClassesI18N['contributions.title']) },
-  {
-    value: 'ACTIVITIES_UPDATES',
-    label: intl.formatMessage(ActivityClassesI18N['activitiesUpdates.title']),
-  },
-];
+import { Timeline } from './Timeline';
 
 const Home = ({ accountSlug }: DashboardSectionProps) => {
   const router = useRouter();
-  const intl = useIntl();
   const [showFeedbackModal, setShowFeedbackModal] = React.useState(false);
-  const filterOptions = React.useMemo(() => getFilterOptions(intl), [intl]);
-  const [filters, setFilters] = React.useState(filterOptions);
-  const [isTimelineBeingGenerated, setIsTimelineBeingGenerated] = React.useState(false);
-  const [openExpenseLegacyId, setOpenExpenseLegacyId] = React.useState<number | null>(null);
   const slug = router.query?.as || accountSlug;
-  const { data, loading, error, fetchMore, refetch } = useQuery(workspaceHomeQuery, {
-    variables: { slug, limit: PAGE_SIZE, classes: flatten(filters.map(f => f.value.split(','))) },
-    context: API_V2_CONTEXT,
-    notifyOnNetworkStatusChange: true,
-  });
-
-  const activities: WorkspaceHomeQuery['account']['feed'] = data?.account.feed || [];
-  const canViewMore = activities.length >= PAGE_SIZE && activities.length % PAGE_SIZE === 0;
-  React.useEffect(() => {
-    if (error?.graphQLErrors?.[0]?.extensions?.code === 'ContentNotReady') {
-      setIsTimelineBeingGenerated(true);
-      setTimeout(() => refetch(), 1000);
-    } else if (data?.account?.feed) {
-      setIsTimelineBeingGenerated(false);
-    }
-  }, [error, data]);
 
   return (
     <div className="flex max-w-screen-lg flex-col-reverse xl:flex-row">
-      <div className="flex-1">
+      <div className="flex flex-1 flex-col gap-4">
         <DashboardHeader
           title={<FormattedMessage id="AdminPanel.Menu.Overview" defaultMessage="Overview" />}
           description={
@@ -83,84 +32,7 @@ const Home = ({ accountSlug }: DashboardSectionProps) => {
             />
           }
         />
-
-        <Flex flexDirection="column" mt="36px">
-          <Flex
-            justifyContent="space-between"
-            alignItems={[null, 'center']}
-            flexDirection={['column', 'row']}
-            gap="8px"
-          >
-            <H2 fontSize="20px" lineHeight="28px" fontWeight="700">
-              <FormattedMessage id="Dashboard.Home.ActivityHeader" defaultMessage="Recent activity" />
-            </H2>
-
-            <StyledSelectFilter
-              intl={intl}
-              inputId="activity-filter"
-              isClearable={false}
-              onChange={setFilters}
-              options={filterOptions}
-              components={REACT_SELECT_COMPONENT_OVERRIDE}
-              value={filters}
-              closeMenuOnSelect={false}
-              hideSelectedOptions={false}
-              isMulti
-              maxWidth={['100%', 300]}
-              minWidth={150}
-              styles={{
-                control: { flexWrap: 'nowrap' },
-              }}
-            />
-          </Flex>
-          <div className="mt-4 space-y-4">
-            {error && !isTimelineBeingGenerated ? (
-              <MessageBoxGraphqlError error={error} />
-            ) : isTimelineBeingGenerated || (!activities.length && loading) ? (
-              <React.Fragment>
-                {isTimelineBeingGenerated && (
-                  <MessageBox type="info" withIcon mb="24px">
-                    <FormattedMessage defaultMessage="Generating activity timeline..." />
-                  </MessageBox>
-                )}
-                <TimelineItem />
-                <TimelineItem />
-                <TimelineItem />
-                <TimelineItem />
-                <TimelineItem />
-              </React.Fragment>
-            ) : !activities.length ? (
-              <MessageBox type="info" withIcon>
-                <FormattedMessage defaultMessage="No activity yet" />
-              </MessageBox>
-            ) : (
-              activities.map(activity => (
-                <TimelineItem key={activity.id} activity={activity} openExpense={id => setOpenExpenseLegacyId(id)} />
-              ))
-            )}
-          </div>
-          {canViewMore && (
-            <StyledButton
-              mt={4}
-              width="100%"
-              buttonSize="small"
-              loading={loading}
-              onClick={() =>
-                fetchMore({
-                  variables: { dateTo: activities[activities.length - 1].createdAt },
-                  updateQuery: (prevResult, { fetchMoreResult }) => {
-                    const account = fetchMoreResult?.account;
-                    account.feed = [...prevResult.account.feed, ...account.feed];
-                    return { account };
-                  },
-                })
-              }
-            >
-              <FormattedMessage defaultMessage="View more" />
-            </StyledButton>
-          )}
-          <ExpenseDrawer openExpenseLegacyId={openExpenseLegacyId} handleClose={() => setOpenExpenseLegacyId(null)} />
-        </Flex>
+        <Timeline accountSlug={slug} withFilter />
       </div>
       <div className="xl:ml-8 xl:w-64">
         <DismissibleMessage messageId={HELP_MESSAGE.WELCOME_TO_DASHBOARD}>
