+++ lib/graphql/types/v2/gql.d.ts
@@ -190,11 +190,11 @@ export function gql(source: "\n  query HostedVirtualCards(\n    $slug: String\n 
   /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function gql(source: "\n  query WorkspaceHome($slug: String!, $limit: Int, $dateTo: DateTime, $classes: [ActivityClassType!]) {\n    account(slug: $slug) {\n      id\n      feed(limit: $limit, dateTo: $dateTo, classes: $classes) {\n        id\n        createdAt\n        type\n        data\n        isSystem\n        fromAccount {\n          id\n          name\n          slug\n          type\n          isIncognito\n          imageUrl(height: 48)\n          ... on Individual {\n            isGuest\n          }\n        }\n        host {\n          id\n          name\n          slug\n          type\n        }\n        account {\n          id\n          name\n          slug\n          type\n          isIncognito\n          imageUrl(height: 48)\n          ... on Individual {\n            isGuest\n          }\n          ... on AccountWithParent {\n            parent {\n              id\n              slug\n              name\n              type\n            }\n          }\n        }\n        expense {\n          id\n          legacyId\n          description\n          account {\n            id\n            name\n            type\n            slug\n            ... on AccountWithParent {\n              parent {\n                id\n                slug\n              }\n            }\n          }\n        }\n        order {\n          id\n          legacyId\n          description\n          toAccount {\n            id\n            name\n            slug\n            ... on AccountWithParent {\n              parent {\n                id\n                slug\n              }\n            }\n          }\n        }\n        update {\n          id\n          legacyId\n          title\n          summary\n          slug\n        }\n        individual {\n          id\n          slug\n          name\n          type\n          imageUrl(height: 48)\n          isIncognito\n        }\n      }\n    }\n  }\n"): typeof import('./graphql').WorkspaceHomeDocument;
+export function gql(source: "\n  query Timeline($slug: String!, $limit: Int, $dateTo: DateTime, $classes: [ActivityClassType!]) {\n    account(slug: $slug) {\n      id\n      feed(limit: $limit, dateTo: $dateTo, classes: $classes) {\n        id\n        createdAt\n        type\n        data\n        isSystem\n        fromAccount {\n          id\n          name\n          slug\n          type\n          isIncognito\n          imageUrl(height: 48)\n          ... on Individual {\n            isGuest\n          }\n        }\n        host {\n          id\n          name\n          slug\n          type\n        }\n        account {\n          id\n          name\n          slug\n          type\n          isIncognito\n          imageUrl(height: 48)\n          ... on Individual {\n            isGuest\n          }\n          ... on AccountWithParent {\n            parent {\n              id\n              slug\n              name\n              type\n            }\n          }\n        }\n        expense {\n          id\n          legacyId\n          description\n          account {\n            id\n            name\n            type\n            slug\n            ... on AccountWithParent {\n              parent {\n                id\n                slug\n              }\n            }\n          }\n        }\n        order {\n          id\n          legacyId\n          description\n          toAccount {\n            id\n            name\n            slug\n            ... on AccountWithParent {\n              parent {\n                id\n                slug\n              }\n            }\n          }\n        }\n        update {\n          id\n          legacyId\n          title\n          summary\n          slug\n        }\n        individual {\n          id\n          slug\n          name\n          type\n          imageUrl(height: 48)\n          isIncognito\n        }\n      }\n    }\n  }\n"): typeof import('./graphql').TimelineDocument;
   /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function gql(source: "\n  query CollectiveOverview(\n    $slug: String!\n    $dateFrom: DateTime\n    $dateTo: DateTime\n    $compareFrom: DateTime\n    $compareTo: DateTime\n    $includeComparison: Boolean!\n    $includeChildren: Boolean!\n  ) {\n    account(slug: $slug) {\n      name\n      imageUrl\n      slug\n      ...AccountHoverCardFields\n      stats {\n        totalBalance: balance(includeChildren: $includeChildren) {\n          currency\n          valueInCents\n        }\n        balance(includeChildren: false) @include(if: $includeChildren) {\n          currency\n          valueInCents\n        }\n        spent: totalAmountSpent(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n          currency\n          valueInCents\n        }\n        spentComparison: totalAmountSpent(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n        received: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          net: true\n        ) {\n          currency\n          valueInCents\n        }\n        receivedComparison: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n        contributionsCount: contributionsCount(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo)\n        contributionsCountComparison: contributionsCount(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n        ) @include(if: $includeComparison)\n      }\n\n      # newOrders: orders(dateFrom: $dateFrom, dateTo: $dateTo) {\n      #   totalCount\n      #   nodes {\n      #     id\n      #     createdAt\n      #     amount {\n      #       currency\n      #       valueInCents\n      #     }\n      #     fromAccount {\n      #       id\n      #       slug\n      #       name\n      #       imageUrl\n      #     }\n      #     status\n      #     frequency\n      #   }\n      # }\n      pendingExpenses: expenses(\n        limit: 0\n        status: PENDING\n        direction: RECEIVED\n        includeChildrenExpenses: $includeChildren\n      ) {\n        totalCount\n      }\n\n      childrenAccounts {\n        totalCount\n        nodes {\n          id\n          name\n          slug\n          imageUrl\n          isArchived\n          ...AccountHoverCardFields\n          ... on Event {\n            startsAt\n            endsAt\n          }\n          stats {\n            balance {\n              currency\n              valueInCents\n            }\n          }\n        }\n      }\n    }\n  }\n  \n"): typeof import('./graphql').CollectiveOverviewDocument;
+export function gql(source: "\n  query CollectiveOverview(\n    $slug: String!\n    $dateFrom: DateTime\n    $dateTo: DateTime\n    $compareFrom: DateTime\n    $compareTo: DateTime\n    $includeComparison: Boolean!\n    $includeChildren: Boolean!\n  ) {\n    account(slug: $slug) {\n      id\n      name\n      imageUrl\n      slug\n      ...AccountHoverCardFields\n      stats {\n        totalBalance: balance(includeChildren: $includeChildren) {\n          currency\n          valueInCents\n        }\n        balance(includeChildren: false) @include(if: $includeChildren) {\n          currency\n          valueInCents\n        }\n        spent: totalAmountSpent(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo, net: true) {\n          currency\n          valueInCents\n        }\n        spentComparison: totalAmountSpent(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n        received: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $dateFrom\n          dateTo: $dateTo\n          net: true\n        ) {\n          currency\n          valueInCents\n        }\n        receivedComparison: totalAmountReceived(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n          net: true\n        ) @include(if: $includeComparison) {\n          currency\n          valueInCents\n        }\n        contributionsCount: contributionsCount(includeChildren: $includeChildren, dateFrom: $dateFrom, dateTo: $dateTo)\n        contributionsCountComparison: contributionsCount(\n          includeChildren: $includeChildren\n          dateFrom: $compareFrom\n          dateTo: $compareTo\n        ) @include(if: $includeComparison)\n      }\n\n      # newOrders: orders(dateFrom: $dateFrom, dateTo: $dateTo) {\n      #   totalCount\n      #   nodes {\n      #     id\n      #     createdAt\n      #     amount {\n      #       currency\n      #       valueInCents\n      #     }\n      #     fromAccount {\n      #       id\n      #       slug\n      #       name\n      #       imageUrl\n      #     }\n      #     status\n      #     frequency\n      #   }\n      # }\n      pendingExpenses: expenses(\n        limit: 0\n        status: PENDING\n        direction: RECEIVED\n        includeChildrenExpenses: $includeChildren\n      ) {\n        totalCount\n      }\n\n      childrenAccounts {\n        totalCount\n        nodes {\n          id\n          name\n          slug\n          imageUrl\n          isArchived\n          ...AccountHoverCardFields\n          ... on Event {\n            startsAt\n            endsAt\n          }\n          stats {\n            balance {\n              currency\n              valueInCents\n            }\n          }\n        }\n      }\n    }\n  }\n  \n"): typeof import('./graphql').CollectiveOverviewDocument;
   /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
