+++ lib/graphql/types/v2/graphql.ts
@@ -10902,7 +10902,7 @@ export type HostedVirtualCardsQueryVariables = Exact<{
 
 export type HostedVirtualCardsQuery = { __typename?: 'Query', host?: { __typename?: 'Host', id: string, legacyId: number, slug: string, name?: string | null, imageUrl?: string | null, currency?: string | null, settings: any, stripe?: { __typename?: 'StripeConnectedAccount', username?: string | null } | null, hostedVirtualCards: { __typename?: 'VirtualCardCollection', totalCount?: number | null, limit?: number | null, offset?: number | null, nodes?: Array<{ __typename?: 'VirtualCard', id?: string | null, name?: string | null, last4?: string | null, data?: any | null, status?: VirtualCardStatus | null, privateData?: any | null, provider?: string | null, spendingLimitAmount?: number | null, spendingLimitInterval?: VirtualCardLimitInterval | null, spendingLimitRenewsOn?: any | null, remainingLimit?: number | null, currency?: Currency | null, createdAt?: any | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, type: AccountType, description?: string | null, isHost: boolean } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, imageUrl?: string | null, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, type: AccountType, description?: string | null, isHost: boolean } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, type: AccountType, description?: string | null, isHost: boolean } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, type: AccountType, description?: string | null, isHost: boolean } | null, assignee?: { __typename?: 'Individual', id: string, name?: string | null, email?: string | null, slug: string, imageUrl?: string | null, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean } | null } | null> | null } } | null };
 
-export type WorkspaceHomeQueryVariables = Exact<{
+export type TimelineQueryVariables = Exact<{
   slug: Scalars['String'];
   limit?: InputMaybe<Scalars['Int']>;
   dateTo?: InputMaybe<Scalars['DateTime']>;
@@ -10910,7 +10910,7 @@ export type WorkspaceHomeQueryVariables = Exact<{
 }>;
 
 
-export type WorkspaceHomeQuery = { __typename?: 'Query', account?: { __typename?: 'Bot', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Collective', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Event', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Fund', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Host', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Individual', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Organization', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Project', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Vendor', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | null };
+export type TimelineQuery = { __typename?: 'Query', account?: { __typename?: 'Bot', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Collective', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Event', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Fund', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Host', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Individual', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Organization', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Project', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | { __typename?: 'Vendor', id: string, feed?: Array<{ __typename?: 'Activity', id: string, createdAt: any, type: ActivityType, data: any, isSystem: boolean, fromAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, host?: { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType } | null, account?: { __typename?: 'Bot', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Event', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Individual', isGuest: boolean, id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null, parent?: { __typename?: 'Bot', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Collective', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Event', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Fund', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Host', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Organization', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Project', id: string, slug: string, name?: string | null, type: AccountType } | { __typename?: 'Vendor', id: string, slug: string, name?: string | null, type: AccountType } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, type: AccountType, isIncognito: boolean, imageUrl?: string | null } | null, expense?: { __typename?: 'Expense', id: string, legacyId: number, description: string, account: { __typename?: 'Bot', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Event', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Host', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, type: AccountType, slug: string } | { __typename?: 'Project', id: string, name?: string | null, type: AccountType, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, type: AccountType, slug: string } } | null, order?: { __typename?: 'Order', id: string, legacyId: number, description?: string | null, toAccount?: { __typename?: 'Bot', id: string, name?: string | null, slug: string } | { __typename?: 'Collective', id: string, name?: string | null, slug: string } | { __typename?: 'Event', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string } | { __typename?: 'Host', id: string, name?: string | null, slug: string } | { __typename?: 'Individual', id: string, name?: string | null, slug: string } | { __typename?: 'Organization', id: string, name?: string | null, slug: string } | { __typename?: 'Project', id: string, name?: string | null, slug: string, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string } | null } | null, update?: { __typename?: 'Update', id: string, legacyId?: number | null, title: string, summary?: string | null, slug: string } | null, individual?: { __typename?: 'Individual', id: string, slug: string, name?: string | null, type: AccountType, imageUrl?: string | null, isIncognito: boolean } | null } | null> | null } | null };
 
 export type CollectiveOverviewQueryVariables = Exact<{
   slug: Scalars['String'];
@@ -10923,7 +10923,7 @@ export type CollectiveOverviewQueryVariables = Exact<{
 }>;
 
 
-export type CollectiveOverviewQuery = { __typename?: 'Query', account?: { __typename?: 'Bot', name?: string | null, imageUrl?: string | null, slug: string, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Collective', name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', name?: string | null, imageUrl?: string | null, slug: string, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Individual', name?: string | null, imageUrl?: string | null, slug: string, isGuest: boolean, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Organization', name?: string | null, imageUrl?: string | null, slug: string, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Project', name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', name?: string | null, imageUrl?: string | null, slug: string, id: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | null };
+export type CollectiveOverviewQuery = { __typename?: 'Query', account?: { __typename?: 'Bot', id: string, name?: string | null, imageUrl?: string | null, slug: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Collective', id: string, name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', id: string, name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, imageUrl?: string | null, slug: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Individual', id: string, name?: string | null, imageUrl?: string | null, slug: string, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Organization', id: string, name?: string | null, imageUrl?: string | null, slug: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | { __typename?: 'Project', id: string, name?: string | null, imageUrl?: string | null, slug: string, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null }, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, imageUrl?: string | null, slug: string, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', contributionsCount: number, contributionsCountComparison: number, totalBalance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, balance?: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spent: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, spentComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, received: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null }, receivedComparison: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, pendingExpenses: { __typename?: 'ExpenseCollection', totalCount?: number | null }, childrenAccounts: { __typename?: 'AccountCollection', totalCount?: number | null, nodes?: Array<{ __typename?: 'Bot', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Collective', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Event', startsAt?: any | null, endsAt?: any | null, id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Fund', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null } | { __typename?: 'Host', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Individual', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, isGuest: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Organization', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | { __typename?: 'Project', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, approvedAt?: any | null, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null, host?: { __typename?: 'Host', id: string, slug: string } | null, parent?: { __typename?: 'Bot', id: string, slug: string } | { __typename?: 'Collective', id: string, slug: string } | { __typename?: 'Event', id: string, slug: string } | { __typename?: 'Fund', id: string, slug: string } | { __typename?: 'Host', id: string, slug: string } | { __typename?: 'Individual', id: string, slug: string } | { __typename?: 'Organization', id: string, slug: string } | { __typename?: 'Project', id: string, slug: string } | { __typename?: 'Vendor', id: string, slug: string } | null } | { __typename?: 'Vendor', id: string, name?: string | null, slug: string, imageUrl?: string | null, isArchived: boolean, type: AccountType, description?: string | null, isHost: boolean, stats?: { __typename?: 'AccountStats', balance: { __typename?: 'Amount', currency?: Currency | null, valueInCents?: number | null } } | null } | null> | null } } | null };
 
 export type TransactionsMetaDataQueryVariables = Exact<{
   slug: Scalars['String'];
@@ -11753,8 +11753,8 @@ export const HostVirtualCardRequestsDocument = {"kind":"Document","definitions":
 export const HostedVirtualCardAccountsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"HostedVirtualCardAccounts"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"hostedVirtualCardCollectives"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"500"}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}},{"kind":"Field","name":{"kind":"Name","value":"limit"}},{"kind":"Field","name":{"kind":"Name","value":"offset"}},{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AccountHoverCardFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Account"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isHost"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"approvedAt"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]} as unknown as DocumentNode<HostedVirtualCardAccountsQuery, HostedVirtualCardAccountsQueryVariables>;
 export const HostedVirtualCardsMetaDataDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"HostedVirtualCardsMetaData"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}}]}}]}}]} as unknown as DocumentNode<HostedVirtualCardsMetaDataQuery, HostedVirtualCardsMetaDataQueryVariables>;
 export const HostedVirtualCardsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"HostedVirtualCards"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"limit"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"offset"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"account"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"AccountReferenceInput"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"status"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"VirtualCardStatus"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"amountFrom"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"AmountInput"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"amountTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"AmountInput"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"hasMissingReceipts"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"searchTerm"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"orderBy"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"ChronologicalOrderInput"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"settings"}},{"kind":"Field","name":{"kind":"Name","value":"stripe"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"username"}}]}},{"kind":"Field","name":{"kind":"Name","value":"hostedVirtualCards"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"Variable","name":{"kind":"Name","value":"limit"}}},{"kind":"Argument","name":{"kind":"Name","value":"offset"},"value":{"kind":"Variable","name":{"kind":"Name","value":"offset"}}},{"kind":"Argument","name":{"kind":"Name","value":"collectiveAccountIds"},"value":{"kind":"Variable","name":{"kind":"Name","value":"account"}}},{"kind":"Argument","name":{"kind":"Name","value":"status"},"value":{"kind":"Variable","name":{"kind":"Name","value":"status"}}},{"kind":"Argument","name":{"kind":"Name","value":"withExpensesDateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"withExpensesDateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"spentAmountFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"amountFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"spentAmountTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"amountTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"hasMissingReceipts"},"value":{"kind":"Variable","name":{"kind":"Name","value":"hasMissingReceipts"}}},{"kind":"Argument","name":{"kind":"Name","value":"searchTerm"},"value":{"kind":"Variable","name":{"kind":"Name","value":"searchTerm"}}},{"kind":"Argument","name":{"kind":"Name","value":"orderBy"},"value":{"kind":"Variable","name":{"kind":"Name","value":"orderBy"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}},{"kind":"Field","name":{"kind":"Name","value":"limit"}},{"kind":"Field","name":{"kind":"Name","value":"offset"}},{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"last4"}},{"kind":"Field","name":{"kind":"Name","value":"data"}},{"kind":"Field","name":{"kind":"Name","value":"status"}},{"kind":"Field","name":{"kind":"Name","value":"privateData"}},{"kind":"Field","name":{"kind":"Name","value":"provider"}},{"kind":"Field","name":{"kind":"Name","value":"spendingLimitAmount"}},{"kind":"Field","name":{"kind":"Name","value":"spendingLimitInterval"}},{"kind":"Field","name":{"kind":"Name","value":"spendingLimitRenewsOn"}},{"kind":"Field","name":{"kind":"Name","value":"remainingLimit"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"account"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}}]}},{"kind":"Field","name":{"kind":"Name","value":"assignee"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"email"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AccountHoverCardFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Account"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isHost"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"approvedAt"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]} as unknown as DocumentNode<HostedVirtualCardsQuery, HostedVirtualCardsQueryVariables>;
-export const WorkspaceHomeDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"WorkspaceHome"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"limit"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"classes"}},"type":{"kind":"ListType","type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ActivityClassType"}}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"account"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"feed"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"Variable","name":{"kind":"Name","value":"limit"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"classes"},"value":{"kind":"Variable","name":{"kind":"Name","value":"classes"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"data"}},{"kind":"Field","name":{"kind":"Name","value":"isSystem"}},{"kind":"Field","name":{"kind":"Name","value":"fromAccount"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"isIncognito"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"48"}}]},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}}]}},{"kind":"Field","name":{"kind":"Name","value":"account"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"isIncognito"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"48"}}]},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}}]}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"expense"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"account"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"order"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"toAccount"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"update"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"summary"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"individual"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"48"}}]},{"kind":"Field","name":{"kind":"Name","value":"isIncognito"}}]}}]}}]}}]}}]} as unknown as DocumentNode<WorkspaceHomeQuery, WorkspaceHomeQueryVariables>;
-export const CollectiveOverviewDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"CollectiveOverview"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"account"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}},{"kind":"Field","name":{"kind":"Name","value":"stats"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"totalBalance"},"name":{"kind":"Name","value":"balance"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","name":{"kind":"Name","value":"balance"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"BooleanValue","value":false}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"spent"},"name":{"kind":"Name","value":"totalAmountSpent"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"spentComparison"},"name":{"kind":"Name","value":"totalAmountSpent"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"received"},"name":{"kind":"Name","value":"totalAmountReceived"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"receivedComparison"},"name":{"kind":"Name","value":"totalAmountReceived"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"contributionsCount"},"name":{"kind":"Name","value":"contributionsCount"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}}]},{"kind":"Field","alias":{"kind":"Name","value":"contributionsCountComparison"},"name":{"kind":"Name","value":"contributionsCount"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}}}]}]}]}},{"kind":"Field","alias":{"kind":"Name","value":"pendingExpenses"},"name":{"kind":"Name","value":"expenses"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"0"}},{"kind":"Argument","name":{"kind":"Name","value":"status"},"value":{"kind":"EnumValue","value":"PENDING"}},{"kind":"Argument","name":{"kind":"Name","value":"direction"},"value":{"kind":"EnumValue","value":"RECEIVED"}},{"kind":"Argument","name":{"kind":"Name","value":"includeChildrenExpenses"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}}]}},{"kind":"Field","name":{"kind":"Name","value":"childrenAccounts"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}},{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isArchived"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Event"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"startsAt"}},{"kind":"Field","name":{"kind":"Name","value":"endsAt"}}]}},{"kind":"Field","name":{"kind":"Name","value":"stats"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"balance"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AccountHoverCardFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Account"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isHost"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"approvedAt"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]} as unknown as DocumentNode<CollectiveOverviewQuery, CollectiveOverviewQueryVariables>;
+export const TimelineDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Timeline"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"limit"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"classes"}},"type":{"kind":"ListType","type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ActivityClassType"}}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"account"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"feed"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"Variable","name":{"kind":"Name","value":"limit"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"classes"},"value":{"kind":"Variable","name":{"kind":"Name","value":"classes"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"data"}},{"kind":"Field","name":{"kind":"Name","value":"isSystem"}},{"kind":"Field","name":{"kind":"Name","value":"fromAccount"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"isIncognito"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"48"}}]},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}}]}},{"kind":"Field","name":{"kind":"Name","value":"account"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"isIncognito"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"48"}}]},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}}]}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"expense"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"account"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"order"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"toAccount"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"update"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"summary"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"individual"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"48"}}]},{"kind":"Field","name":{"kind":"Name","value":"isIncognito"}}]}}]}}]}}]}}]} as unknown as DocumentNode<TimelineQuery, TimelineQueryVariables>;
+export const CollectiveOverviewDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"CollectiveOverview"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"DateTime"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"account"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}},{"kind":"Field","name":{"kind":"Name","value":"stats"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"totalBalance"},"name":{"kind":"Name","value":"balance"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","name":{"kind":"Name","value":"balance"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"BooleanValue","value":false}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"spent"},"name":{"kind":"Name","value":"totalAmountSpent"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"spentComparison"},"name":{"kind":"Name","value":"totalAmountSpent"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"received"},"name":{"kind":"Name","value":"totalAmountReceived"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"receivedComparison"},"name":{"kind":"Name","value":"totalAmountReceived"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}}},{"kind":"Argument","name":{"kind":"Name","value":"net"},"value":{"kind":"BooleanValue","value":true}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"contributionsCount"},"name":{"kind":"Name","value":"contributionsCount"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"dateTo"}}}]},{"kind":"Field","alias":{"kind":"Name","value":"contributionsCountComparison"},"name":{"kind":"Name","value":"contributionsCount"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"includeChildren"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateFrom"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareFrom"}}},{"kind":"Argument","name":{"kind":"Name","value":"dateTo"},"value":{"kind":"Variable","name":{"kind":"Name","value":"compareTo"}}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeComparison"}}}]}]}]}},{"kind":"Field","alias":{"kind":"Name","value":"pendingExpenses"},"name":{"kind":"Name","value":"expenses"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"0"}},{"kind":"Argument","name":{"kind":"Name","value":"status"},"value":{"kind":"EnumValue","value":"PENDING"}},{"kind":"Argument","name":{"kind":"Name","value":"direction"},"value":{"kind":"EnumValue","value":"RECEIVED"}},{"kind":"Argument","name":{"kind":"Name","value":"includeChildrenExpenses"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includeChildren"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}}]}},{"kind":"Field","name":{"kind":"Name","value":"childrenAccounts"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}},{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isArchived"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Event"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"startsAt"}},{"kind":"Field","name":{"kind":"Name","value":"endsAt"}}]}},{"kind":"Field","name":{"kind":"Name","value":"stats"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"balance"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"valueInCents"}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AccountHoverCardFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Account"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isHost"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"approvedAt"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}}]} as unknown as DocumentNode<CollectiveOverviewQuery, CollectiveOverviewQueryVariables>;
 export const TransactionsMetaDataDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"TransactionsMetaData"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"transactions"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"account"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}]}},{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"0"}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"paymentMethodTypes"}},{"kind":"Field","name":{"kind":"Name","value":"kinds"}}]}},{"kind":"Field","name":{"kind":"Name","value":"account"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"legacyId"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}}]}}]}}]} as unknown as DocumentNode<TransactionsMetaDataQuery, TransactionsMetaDataQueryVariables>;
 export const DashboardVendorsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"DashboardVendors"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"searchTerm"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"isArchived"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"includePotentialVendors"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"limit"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"offset"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"account"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"slug"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Organization"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"legalName"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"expensePolicy"}},{"kind":"Field","name":{"kind":"Name","value":"settings"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"features"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"MULTI_CURRENCY_EXPENSES"}}]}},{"kind":"Field","name":{"kind":"Name","value":"location"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"address"}},{"kind":"Field","name":{"kind":"Name","value":"country"}}]}},{"kind":"Field","name":{"kind":"Name","value":"transferwise"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"availableCurrencies"}}]}},{"kind":"Field","name":{"kind":"Name","value":"supportedPayoutMethods"}},{"kind":"Field","name":{"kind":"Name","value":"isTrustedHost"}},{"kind":"Field","name":{"kind":"Name","value":"vendors"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"searchTerm"},"value":{"kind":"Variable","name":{"kind":"Name","value":"searchTerm"}}},{"kind":"Argument","name":{"kind":"Name","value":"isArchived"},"value":{"kind":"Variable","name":{"kind":"Name","value":"isArchived"}}},{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"Variable","name":{"kind":"Name","value":"limit"}}},{"kind":"Argument","name":{"kind":"Name","value":"offset"},"value":{"kind":"Variable","name":{"kind":"Name","value":"offset"}}}],"directives":[{"kind":"Directive","name":{"kind":"Name","value":"skip"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includePotentialVendors"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}},{"kind":"Field","name":{"kind":"Name","value":"offset"}},{"kind":"Field","name":{"kind":"Name","value":"limit"}},{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"VendorFields"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"potentialVendors"},"directives":[{"kind":"Directive","name":{"kind":"Name","value":"include"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"if"},"value":{"kind":"Variable","name":{"kind":"Name","value":"includePotentialVendors"}}}]}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"tags"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"96"}}]},{"kind":"Field","name":{"kind":"Name","value":"isArchived"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AccountHoverCardFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Account"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"Field","name":{"kind":"Name","value":"isHost"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Individual"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isGuest"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"host"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}},{"kind":"Field","name":{"kind":"Name","value":"approvedAt"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithParent"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"parent"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"VendorFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Vendor"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"tags"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"height"},"value":{"kind":"IntValue","value":"96"}}]},{"kind":"Field","name":{"kind":"Name","value":"hasImage"}},{"kind":"Field","name":{"kind":"Name","value":"isArchived"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"location"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"address"}},{"kind":"Field","name":{"kind":"Name","value":"country"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"structured"}}]}},{"kind":"Field","name":{"kind":"Name","value":"createdByAccount"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"imageUrl"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"AccountHoverCardFields"}}]}},{"kind":"Field","name":{"kind":"Name","value":"vendorInfo"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"contact"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"email"}}]}},{"kind":"Field","name":{"kind":"Name","value":"taxFormUrl"}},{"kind":"Field","name":{"kind":"Name","value":"taxFormRequired"}},{"kind":"Field","name":{"kind":"Name","value":"taxType"}},{"kind":"Field","name":{"kind":"Name","value":"taxId"}},{"kind":"Field","name":{"kind":"Name","value":"notes"}}]}},{"kind":"Field","name":{"kind":"Name","value":"payoutMethods"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"type"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"data"}}]}},{"kind":"Field","name":{"kind":"Name","value":"orders"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"filter"},"value":{"kind":"EnumValue","value":"OUTGOING"}},{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"1"}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}}]}},{"kind":"Field","name":{"kind":"Name","value":"expenses"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"status"},"value":{"kind":"EnumValue","value":"PAID"}},{"kind":"Argument","name":{"kind":"Name","value":"direction"},"value":{"kind":"EnumValue","value":"SUBMITTED"}},{"kind":"Argument","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"1"}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"totalCount"}}]}}]}}]} as unknown as DocumentNode<DashboardVendorsQuery, DashboardVendorsQueryVariables>;
 export const EditAccountFeeDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"EditAccountFee"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"account"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"AccountReferenceInput"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"hostFeePercent"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"isCustomFee"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Boolean"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"editAccountFeeStructure"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"account"},"value":{"kind":"Variable","name":{"kind":"Name","value":"account"}}},{"kind":"Argument","name":{"kind":"Name","value":"hostFeePercent"},"value":{"kind":"Variable","name":{"kind":"Name","value":"hostFeePercent"}}},{"kind":"Argument","name":{"kind":"Name","value":"isCustomFee"},"value":{"kind":"Variable","name":{"kind":"Name","value":"isCustomFee"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"hostFeesStructure"}},{"kind":"Field","name":{"kind":"Name","value":"hostFeePercent"}}]}},{"kind":"Field","name":{"kind":"Name","value":"childrenAccounts"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"nodes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"AccountWithHost"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"hostFeesStructure"}},{"kind":"Field","name":{"kind":"Name","value":"hostFeePercent"}}]}}]}}]}}]}}]}}]} as unknown as DocumentNode<EditAccountFeeMutation, EditAccountFeeMutationVariables>;
